requirements.txt
opencv_python
pygame 
ultralytics
pyttsx3
gtts
firebase_admin
fastapi
uvicorn 
opencv-python 
websockets
mpu6050-raspberrypi 
pynmea2